+++ backend/lcfs/web/api/compliance_report/schema.py
@@ -162,11 +162,11 @@ class ComplianceReportSummaryRowSchema(BaseSchema):
     line: Optional[str] = ""
     description: Optional[str] = ""
     field: Optional[str] = ""
-    gasoline: Optional[float] = 0
-    diesel: Optional[float] = 0
-    jet_fuel: Optional[float] = 0
-    value: Optional[float] = 0
-    total_value: Optional[float] = 0
+    gasoline: Optional[int] = 0
+    diesel: Optional[int] = 0
+    jet_fuel: Optional[int] = 0
+    value: Optional[int] = 0
+    total_value: Optional[int] = 0
     format: Optional[str] = ""
